Refactor error handling and add method comments

+++ src/main/java/org/unibl/etf/ip2024/controllers/UserProgramController.java
@@ -19,6 +19,7 @@ public class UserProgramController {
 
     private final UserProgramService userProgramService;
 
+    // Endpoint for getting user programs
     @GetMapping
     public ResponseEntity<Page<FitnessProgramListResponse>> getUserPrograms(
             Principal principal,
@@ -29,16 +30,24 @@ public class UserProgramController {
         return ResponseEntity.ok(userPrograms);
     }
 
+    // Endpoint for creating user program
     @PostMapping("/{programId}")
     public ResponseEntity<UserProgramResponse> createUserProgram(
             Principal principal,
             @PathVariable Integer programId) {
+        if (programId == null || programId <= 0) {
+            return ResponseEntity.badRequest().build();
+        }
         UserProgramResponse userProgram = userProgramService.createUserProgram(principal, programId);
         return ResponseEntity.ok(userProgram);
     }
 
+    // Endpoint for deleting user program
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePurchasedProgram(@PathVariable("id") Integer programId, Principal principal) {
+        if (programId == null || programId <= 0) {
+            return ResponseEntity.badRequest().build();
+        }
         userProgramService.deleteUserProgram(principal, programId);
         return ResponseEntity.noContent().build(); // Status 204 No Content
     }
